damage system finished +++++++
damage system working, both the player and the enemy die, take damage, animate at the right moment, receive and inflict knockback effect and freeze a for a little bit when hit. there is only a very small problem on the enemy local scale that is flipped but it's something that i can do in a min tomorrow, now i'm to lazy and mostly tired.