Working controls canvas & Testing
Can nav from the main menu, can return to main menu, also displays on death as a "git gud" to the player. The one that's displaying within the Game Over screen might need a lil tweaking, I feel it's slightly too large.

I'm wondering about the general size of the ui for the player when they're ingame at the moment

I went through most of the menu loop, it's not crashing anything my end atm which is positve